Migrated phone number validators and normalizers
 * Rather than applying validators/normalizers that perform advanced formatting at the field level, they're applied at the form level.
 * Added a `maxLengthNormalizer` that simply takes a length and returns a function that returns true if the length is less than the specified value.
